Organize code, break up common and utility files, remove unused code
