+++ source/typescript/compiler/states/process_production_chain.ts
@@ -1,12 +1,9 @@
-import {
-    getAccompanyingItems,
-    itemsToProductions,
-    rec_state_prod
-} from "../../util/utilities.js";
-import { SC } from "../../util/skribble.js";
 import { RenderBodyOptions } from "../../types/render_body_options";
-import { renderItemReduction } from "../../util/item_render_functions.js";
-import { Item } from "../../util/item.js";
+
+import { getAccompanyingItems, Item, itemsToProductions } from "../../util/item.js";
+import { renderItemReduction } from "../../util/render_item.js";
+import { SC } from "../../util/skribble.js";
+import { rec_state_prod } from "../../util/global_names.js";
 
 export function processProductionChain(
     code_node: SC,
