+++ source/typescript/compiler/states/yield_nonterm_states.ts
@@ -1,10 +1,12 @@
 
-import { rec_glob_lex_name, isSymAProduction } from "../../util/utilities.js";
-import { SC } from "../../util/skribble.js";
-import { RenderBodyOptions } from "../../types/render_body_options";
-import { yieldStates } from "./yield_states.js";
 import { RecognizerState, TRANSITION_TYPE } from "../../types/recognizer_state.js";
+import { RenderBodyOptions } from "../../types/render_body_options";
+
 import { Item } from "../../util/item.js";
+import { SC } from "../../util/skribble.js";
+import { isSymAProduction } from "../../util/symbol.js";
+import { rec_glob_lex_name } from "../../util/global_names.js";
+import { yieldStates } from "./yield_states.js";
 
 export function* yieldNontermStates(options: RenderBodyOptions): Generator<RecognizerState[], { code: SC, prods: number[]; hash: string; }> {
