+++ source/typescript/types/rd_production_function.ts
@@ -1,8 +1,9 @@
-import { SC } from "../util/skribble";
+import { SC } from "../util/skribble.js";
 export interface RDProductionFunction {
 
     id: number,
     fn?: SC;
     productions?: Set<number>;
     RENDER?: boolean;
+
 }
