+++ source/typescript/types/worker_container.ts
@@ -1,0 +1,7 @@
+import { Worker } from "worker_threads";
+
+export type WorkerContainer = {
+    target: Worker;
+    id: number;
+    READY: boolean;
+};
