+++ source/typescript/util/__global.d.ts
@@ -1,0 +1,2 @@
+declare global {
+}
